AssemblyArt4/requirements.txt
build123d
ocp_vscode
numpy-stl
trimesh
manifold3d
rtree